CASH-1096: Move FundraisingStatus into subdirectory
* CASH-1096 includes a 3rd file in FundraisingStatus directory, prepping for that
