+++ src/components/LoanCards/AdaptiveMicroLoanCard.vue
@@ -54,7 +54,7 @@ import _get from 'lodash/get';
 import _forEach from 'lodash/forEach';
 import _includes from 'lodash/includes';
 import LoanCardImage from '@/components/LoanCards/LoanCardImage';
-import FundraisingStatusMeter from '@/components/LoanCards/FundraisingStatusMeter';
+import FundraisingStatusMeter from '@/components/LoanCards/FundraisingStatus/FundraisingStatusMeter';
 import updateLoanReservation from '@/graphql/mutation/updateLoanReservation.graphql';
 
 export default {
